+++ src/features/login/type.ts
@@ -1,0 +1,5 @@
+export type LoginProps = {
+  email: string
+  pw: string
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void
+}
