+++ src/features/login/ui/index.ts
@@ -1,0 +1,1 @@
+export { LoginInputs } from './LoginInputs'
