+++ src/views/index.ts
@@ -1,1 +1,2 @@
 export { RegistPage } from './register/RegistPage'
+export { LoginPage } from './login/LoginPage'
