+++ src/widgets/index.ts
@@ -1,2 +1,3 @@
 export { RegistForm } from './register/RegistForm.ui'
 export { AuthFooter } from './AuthFooter'
+export { LoginForm } from './login/LoginForm.ui'
